requirements.txt
asgiref==3.9.1
cfgv==3.4.0
distlib==0.4.0
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
filelock==3.18.0
identify==2.6.12
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.2
sqlparse==0.5.3
virtualenv==20.33.0
